refine script for counting terms by decade
