+++ dictionaries/expanded_decades/demographic_1993_2003.txt
@@ -1,0 +1,59 @@
+accountability
+adaptation
+adaptive
+carrying_capacity
+competency_trap
+competitive_intensity
+competitive_pressure
+competitive_pressures
+death_rate
+death_rates
+densities
+density
+ecological
+ecologists
+ecology
+entry_exit
+evolution
+evolutionary
+expansion_contraction
+failure
+failures
+fitness
+founding
+generalist
+generalists
+imprinting
+inertia
+inertial
+legitimacy
+legitimate
+liabilities_newness
+liability_adolescence
+liability_newness
+liability_obsolescence
+liability_senescence
+liability_smallness
+localized
+localized_competition
+mortality
+niche
+niches
+organizational_change
+organizational_form
+organizational_forms
+population
+populations
+red_queen
+resetting_clock
+resistance_change
+resource_partitioning
+selection
+senescence
+size_localized
+specialism
+specialist
+specialization
+survival
+survive
+viability
